requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.3.2
Django==4.2.7
gunicorn==21.2.0
idna==3.4
packaging==23.2
psycopg2-binary==2.9.9
requests==2.31.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.1.0
whitenoise==6.6.0
